+++ cmd/traffic/cmd/agentinit/agent_init.go
@@ -13,8 +13,8 @@ import (
 	"strings"
 
 	"github.com/coreos/go-iptables/iptables"
-	"gopkg.in/yaml.v3"
 	core "k8s.io/api/core/v1"
+	"sigs.k8s.io/yaml"
 
 	"github.com/datawire/dlib/derror"
 	"github.com/datawire/dlib/dlog"
@@ -31,14 +31,13 @@ type config struct {
 }
 
 func loadConfig(ctx context.Context) (*config, error) {
-	cf, err := dos.Open(ctx, filepath.Join(agentconfig.ConfigMountPoint, agentconfig.ConfigFile))
+	bs, err := dos.ReadFile(ctx, filepath.Join(agentconfig.ConfigMountPoint, agentconfig.ConfigFile))
 	if err != nil {
 		return nil, fmt.Errorf("unable to open agent ConfigMap: %w", err)
 	}
-	defer cf.Close()
 
 	c := config{}
-	if err = yaml.NewDecoder(cf).Decode(&c.Sidecar); err != nil {
+	if err = yaml.Unmarshal(bs, &c.Sidecar); err != nil {
 		return nil, fmt.Errorf("unable to decode agent ConfigMap: %w", err)
 	}
 	return &c, nil