+++ cmd/traffic/cmd/manager/internal/mutator/agent_injector_test.go
@@ -723,7 +723,9 @@ func TestTrafficAgentConfigGenerator(t *testing.T) {
 		test := test // pin it
 		ctx := k8sapi.WithK8sInterface(ctx, clientset)
 		t.Run(test.name, func(t *testing.T) {
-			actualConfig, actualErr := generateForPod(t, ctx, test.request, env.GeneratorConfig("docker.io/datawire/tel2:2.6.0"))
+			gc, err := env.GeneratorConfig("docker.io/datawire/tel2:2.6.0")
+			require.NoError(t, err)
+			actualConfig, actualErr := generateForPod(t, ctx, test.request, gc)
 			requireContains(t, actualErr, strings.ReplaceAll(test.expectedError, "<PODNAME>", test.request.Name))
 			if actualConfig == nil {
 				actualConfig = &agentconfig.Sidecar{}
@@ -1615,6 +1617,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 			ctx := dlog.NewTestContext(t, false)
 			ctx = managerutil.WithEnv(ctx, env)
 			ctx = k8sapi.WithK8sInterface(ctx, clientset)
+			ctx = managerutil.WithAgentImageRetriever(ctx, nil)
 			if test.envAdditions != nil {
 				env := managerutil.GetEnv(ctx)
 				newEnv := *env
@@ -1632,14 +1635,16 @@ func TestTrafficAgentInjector(t *testing.T) {
 			var actualErr error
 			cw := NewWatcher("")
 			if test.generateConfig {
+				gc, err := env.GeneratorConfig("docker.io/datawire/tel2:2.6.0")
+				require.NoError(t, err)
 				var ac *agentconfig.Sidecar
-				if ac, actualErr = generateForPod(t, ctx, test.pod, env.GeneratorConfig("docker.io/datawire/tel2:2.6.0")); actualErr == nil {
+				if ac, actualErr = generateForPod(t, ctx, test.pod, gc); actualErr == nil {
 					actualErr = cw.Store(ctx, ac, true)
 				}
 			}
 			if actualErr == nil {
 				request := toAdmissionRequest(podResource, test.pod)
-				a := agentInjector{agentConfigs: cw, agentImage: "docker.io/datawire/tel2:2.6.0"}
+				a := agentInjector{agentConfigs: cw}
 				actualPatch, actualErr = a.inject(ctx, request)
 			}
 			requireContains(t, actualErr, strings.ReplaceAll(test.expectedError, "<PODNAME>", test.pod.Name))