+++ cmd/traffic/cmd/manager/managerutil/systema.go
@@ -5,7 +5,6 @@ import (
 
 	"google.golang.org/grpc"
 
-	"github.com/datawire/dlib/dlog"
 	"github.com/telepresenceio/telepresence/rpc/v2/common"
 	"github.com/telepresenceio/telepresence/rpc/v2/manager"
 	"github.com/telepresenceio/telepresence/rpc/v2/systema"
@@ -18,26 +17,6 @@ import (
 type SystemaCRUDClient interface {
 	systemarpc.SystemACRUDClient
 	a8rcloud.Closeable
-}
-
-func GetAgentImage(ctx context.Context) string {
-	var agentImage string
-	env := GetEnv(ctx)
-
-	// The AgentImage in the container's environment have the highest priority. If it isn't provided,
-	// then we as SystemA for the preferred image. This means that air-gapped environment can declare
-	// their preferred image using the environment variable and refrain from an attempt to contact
-	// SystemA.
-	if env.AgentImage == "" {
-		var err error
-		if agentImage, err = AgentImageFromSystemA(ctx); err != nil {
-			dlog.Errorf(ctx, "unable to get Ambassador Cloud preferred agent image: %v", err)
-		}
-	}
-	if agentImage == "" {
-		agentImage = env.QualifiedAgentImage()
-	}
-	return agentImage
 }
 
 type UnauthdConnProvider struct {